chore: update

+++ packages/contracts/test/dollar/core/UbiquityGovernanceToken.t.sol
@@ -87,8 +87,8 @@ contract UbiquityGovernanceTokenTest is LocalTestHelper {
     }
 
     function testUUPS_InitializedVersion() external {
-        uint expectedVersion = 1;
-        uint baseExpectedVersion = 255;
+        uint256 expectedVersion = 1;
+        uint256 baseExpectedVersion = 255;
 
         UbiquityGovernanceTokenUpgraded newImpl = new UbiquityGovernanceTokenUpgraded();
         UbiquityGovernanceTokenUpgraded newImplT = new UbiquityGovernanceTokenUpgraded();
